+++ src/main/java/org/broadinstitute/hellbender/tools/dataflow/pipelines/ReadsPreprocessingPipeline.java
@@ -19,6 +19,7 @@ import org.broadinstitute.hellbender.dev.DoFnWLog;
 import org.broadinstitute.hellbender.engine.dataflow.*;
 import org.broadinstitute.hellbender.engine.dataflow.datasources.*;
 import org.broadinstitute.hellbender.engine.dataflow.transforms.composite.AddContextDataToRead;
+import org.broadinstitute.hellbender.tools.dataflow.transforms.markduplicates.MarkDuplicates;
 import org.broadinstitute.hellbender.tools.recalibration.RecalibrationArgumentCollection;
 import org.broadinstitute.hellbender.tools.recalibration.RecalibrationTables;
 import org.broadinstitute.hellbender.tools.recalibration.covariates.StandardCovariateList;
@@ -80,7 +81,7 @@ public class ReadsPreprocessingPipeline extends DataflowCommandLineProgram {
         final PCollection<GATKRead> initialReads = readsDataflowSource.getReadPCollection(intervals);
 
         // Apply MarkDuplicates to produce updated GATKReads.
-        final PCollection<GATKRead> markedReads = initialReads.apply(new MarkDuplicatesStub(headerSingleton));
+        final PCollection<GATKRead> markedReads = initialReads.apply(new MarkDuplicates(headerSingleton));
 
         // Load the Variants and the Reference and join them to reads.
         final VariantsDataflowSource variantsDataflowSource = new VariantsDataflowSource(baseRecalibrationKnownVariants, pipeline);
@@ -96,30 +97,6 @@ public class ReadsPreprocessingPipeline extends DataflowCommandLineProgram {
 
         final PCollection<GATKRead> finalReads = markedReads.apply(new ApplyBQSRStub(headerSingleton, mergedRecalibrationReport));
         SmallBamWriter.writeToFile(pipeline, finalReads, readsHeader, output);
-    }
-
-    // NOTE: need a way to ensure that certain tools are guaranteed to have a header -- one option is
-    // an interface with a factory method
-    private static class MarkDuplicatesStub extends PTransform<PCollection<GATKRead>, PCollection<GATKRead>> {
-        private static final long serialVersionUID = 1L;
-
-        private PCollectionView<SAMFileHeader> header;
-
-        public MarkDuplicatesStub( final PCollectionView<SAMFileHeader> header ) {
-            this.header = header;
-        }
-
-        @Override
-        public PCollection<GATKRead> apply( PCollection<GATKRead> input ) {
-            return input.apply(ParDo.named("MarkDuplicates").
-                    of(new DoFnWLog<GATKRead, GATKRead>("MarkDuplicatesStub") {
-                        private static final long serialVersionUID = 1L;
-                        @Override
-                        public void processElement( ProcessContext c ) throws Exception {
-                            c.output(c.element());
-                        }
-                    }).withSideInputs(header));
-        }
     }
 
     private static class BaseRecalibratorStub extends PTransform<PCollection<KV<GATKRead, ReadContextData>>, PCollection<RecalibrationTables>> {
