+++ src/test/java/org/broadinstitute/hellbender/tools/dataflow/transforms/markduplicates/ReadsKeyTest.java
@@ -1,15 +1,14 @@
-package org.broadinstitute.hellbender.tools.dataflow;
+package org.broadinstitute.hellbender.tools.dataflow.transforms.markduplicates;
 
 
 import htsjdk.samtools.SAMFileHeader;
-import org.broadinstitute.hellbender.tools.dataflow.transforms.MarkDuplicatesReadsKey;
 import org.broadinstitute.hellbender.utils.read.ArtificialReadUtils;
 import org.broadinstitute.hellbender.utils.read.GATKRead;
 import org.testng.Assert;
 import org.testng.annotations.DataProvider;
 import org.testng.annotations.Test;
 
-public final class MarkDuplicatesReadsKeyTest {
+public final class ReadsKeyTest {
     @DataProvider(name = "sams")
     public Object[][] sams(){
         final SAMFileHeader header = ArtificialReadUtils.createArtificialSamHeader();
@@ -31,6 +30,6 @@ public final class MarkDuplicatesReadsKeyTest {
 
     @Test(dataProvider = "sams", groups= {"dataflow"})
     public void countBasesTest(final SAMFileHeader header, final GATKRead read, final String expectedKey){
-        final String key = MarkDuplicatesReadsKey.keyForFragment(header, read);
+        final String key = ReadsKey.keyForFragment(header, read);
         Assert.assertEquals(expectedKey, key);
     }}
